Set SocketServer.allow_reuse_address=true
This lets us share the SSDP port happily with other running software.
The kernel forbids us to bind a second time to the same port, if there
is already a socket open. But as this is multicast it is common to have
multiple applications listening to the same multicast address and port.

Setting SO_REUSEADDR on the socket before attempting to bind() will work
happily and make use coexist with other UPnP software on the machine.