Update/Add compatibility info for Freebox set-top-boxes
See issue #220

Secondly, I tested with my Freebox Revolution Player (the set-top-box), which is not the same hardware, and has more powerfull software as well and it works like a charm with the codecs WAV, MP3, FLAC, OGG but not OPUS and L16 (it seems it is refused by the renderer because of the mimetype). The box is detected as a DLNA renderer (even though it also implements AirPlay).

Note that the Freebox Revolution Server (the modem) also has tiny speakers and is also a DLNA renderer, but I didn't test yet.