Remember the device volume and set the volume to 100% during streaming
- Normally the application volume for music is set to about 30%. Since that volume level is not changed when streaming the applications sound to a device and the device itself also has volume level the music can be very silent.
overall_volume = application_volume * device_volume

The idea is to remember the volume level of the device and set the device's volume to 100% while streaming. In that way you eliminate the device's volume and you are just using the computers volume.
overall_volume = application_volume * 1
overall_volume = application_volume

When you terminate the streaming the remembered device volume is getting restored, otherwise you blow up the users ears. Might be problematic when the application crashes ...